D2Ptrs.h :: added function pointer for GetItemLevelRequirement, pointer found by Sheppard JSUnit.cpp and JSUnit.h :: added item.lvlreq to core MPQStats.cpp :: fixed compiler warning bout signed/unsigned mismatch in new function
git-svn-id: https://subversion.assembla.com/svn/d2bs@218 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ trunk/MPQStats.cpp
@@ -53,7 +53,7 @@ DWORD GetBaseTable(INT nBaseStat, INT nClassId)
 			dwMaxEntries = *(DWORD*)(dwMaxEntriesOffset + dwD2MPQTable);
 		else dwMaxEntries = 0xFF;
 
-		if(nClassId < dwMaxEntries)
+		if((DWORD)nClassId < dwMaxEntries)
 		{
 			DWORD dwMultiplicator = BaseStatTable[nBaseStat].pTable[BaseStatTable[nBaseStat].wTableSize-1].dwFieldOffset;
 			DWORD dwTable = nClassId * dwMultiplicator;
